fixed bad casting

+++ dataset.py
@@ -33,7 +33,7 @@ class DomainDataset(Dataset):
         return ret
     
     def compute_items_sequential(self):
-        n_points_per_axis = np.ceil(self.n ** (1/self.dim))
+        n_points_per_axis = int(np.ceil(self.n ** (1/self.dim)))
         xi = []
         for i in range(self.dim):
             s = np.linspace(self.xmin[i], self.xmax[i], num = n_points_per_axis + 1, endpoint=False)[1:]
@@ -42,7 +42,7 @@ class DomainDataset(Dataset):
         return
 
     def compute_items_rand(self):
-        n_points_per_axis = np.ceil(self.n ** (1/self.dim))
+        n_points_per_axis = int(np.ceil(self.n ** (1/self.dim)))
         xi = []
         for i in range(self.dim):
             s = np.random.uniform(low=self.xmin[i], high=self.xmax[i], size=(n_points_per_axis, ))
